requirements/base.txt
Django==4.2.7
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
ddgs>=9.10.0
python-json-logger>=2.0.7
djangorestframework-simplejwt==5.3.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-filter==23.5
dnstwist
python-Wappalyzer
